test(search): see how blocking cpu-futures behave
I think that could already be it, as all we need here is
some operation which runs on the loop and can thus be aborted.

Thanks to the threadpool, operations make progress even without
us waiting for it in the main thread, and thus can be dropped at
any time.