Now with support for deltas per run, all stored by date
This can later be triggered on regular intervals for long-running
criner instances.

I am already looking forward to see this data in a table somewhere,
grouped by date.

Lastly, this really shows we need an abstractions for each type we
store, and it seems quite straightforward to do without code generation.
By now there are enough examples to generalize all of them.