Now with actually storing messages - it's pretty slow, but…
…there is room for improvement when it comes to memory management.
We do one unnecessary allocation, and could safe another one if
we could get the title as reference.

Current performance should even be OK knowing that 'info' should
only rarely be used, and done/fail are rare events too for
the anticipated usage.

If not, some people might be able to contribute a high-performance
ring buffer without unsafe that is exacty what we need.